Implement map and picking address on map. Code improvements.

+++ core/network/src/main/java/com/dvm/network/api/CartApi.kt
@@ -16,7 +16,7 @@ interface CartApi {
     suspend fun checkInput(address: String): AddressResponse
 
     suspend fun checkCoordinates(
-        latitude: Long,
-        longitude: Long
+        latitude: Double,
+        longitude: Double
     ): AddressResponse
 }
